+++ test/Service/ConsoleViewHelperManagerDelegatorFactoryTest.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * @link      http://github.com/zendframework/zend-mvc-console for the canonical source repository
- * @copyright Copyright (c) 2005-2016 Zend Technologies USA Inc. (http://www.zend.com)
+ * @copyright Copyright (c) 2005-2018 Zend Technologies USA Inc. (http://www.zend.com)
  * @license   http://framework.zend.com/license/new-bsd New BSD License
  */
 
@@ -10,7 +10,12 @@ namespace ZendTest\Mvc\Console\Service;
 use Interop\Container\ContainerInterface;
 use PHPUnit_Framework_TestCase as TestCase;
 use Prophecy\Argument;
+use ReflectionMethod;
+use Zend\Mvc\Application;
 use Zend\Mvc\Console\Service\ConsoleViewHelperManagerDelegatorFactory;
+use Zend\Mvc\MvcEvent;
+use Zend\Router\RouteMatch;
+use Zend\Router\RouteStackInterface;
 use Zend\View\HelperPluginManager;
 use Zend\View\Helper;
 
@@ -56,5 +61,31 @@ class ConsoleViewHelperManagerDelegatorFactoryTest extends TestCase
             $this->plugins->reveal(),
             $this->factory->__invoke($this->container, 'ViewHelperManager', $this->callback)
         );
+    }
+
+    public function testCreateUrlHelperFactoryInjectsHelperWithRouterAndRouteMatchWhenPresent()
+    {
+        $container = $this->prophesize(ContainerInterface::class);
+
+        $router = $this->prophesize(RouteStackInterface::class);
+        $container->get('HttpRouter')->will([$router, 'reveal']);
+
+        $routeMatch = $this->prophesize(RouteMatch::class);
+
+        $mvcEvent = $this->prophesize(MvcEvent::class);
+        $mvcEvent->getRouteMatch()->will([$routeMatch, 'reveal']);
+
+        $application = $this->prophesize(Application::class);
+        $application->getMvcEvent()->will([$mvcEvent, 'reveal']);
+        $container->get('Application')->will([$application, 'reveal']);
+
+        $r = new ReflectionMethod($this->factory, 'createUrlHelperFactory');
+        $r->setAccessible(true);
+        $factory = $r->invoke($this->factory, $container->reveal());
+
+        $helper = $factory();
+
+        $this->assertAttributeSame($router->reveal(), 'router', $helper);
+        $this->assertAttributeSame($routeMatch->reveal(), 'routeMatch', $helper);
     }
 }